pyproject.toml
[project]
name = "shop-inventory"
version = "0.1.0"
description = "Local-only inventory management for the Rowan Pantry"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django<5",
    "gunicorn>=23.0.0",
    "pillow>=11.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "qrcode[pil]>=8.0",
    "tqdm>=4.67.1",
    "treepoem>=3.25.0",
    "wait-for-it>=2.3.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.9.0",
    "ruff>=0.7.4",
]
prod = [
    "gunicorn>=23.0.0",
]

[tool.pytest.ini_options]
pythonpath = "src/shop-inventory"
DJANGO_SETTINGS_MODULE = "_core.settings"
python_files = ["test.py", "tests.py", "test_*.py", "*_tests.py",]
addopts = ["--cov=src/shop-inventory", "--cov-config=pyproject.toml",]

[tool.coverage.run]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "manage.py",
    "*/__init__.py",
    "*/apps.py",
    "*/management/commands/*",
    "**/asgi.py",
    "**/wsgi.py"
]
[tool.coverage.report]
skip_covered = true
show_missing = true
exclude_also = [
    "pragma: no cover",
    "def __repr__",
    "if settings.DEBUG",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
